Fixed NullReferenceException when failing to resolve type
Fixes: devdiv.visualstudio.com/DevDiv/_workitems/edit/705725.

GetRuntimeType returns null on a variety of situations under the covers, for example, when it fails load a type because it references another assembly that isn't available to the project.

I repro'd this simply by adding a reference to Microsoft.AspNetCore.Mvc.Abstractions and attempting to expand it in the TypePickerDialog. Before the fix the dialog closes, after the fix it expands with types that didn't fail to load.