Fix crash when adding new setting
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/705752

CreateUniqueName and EqualIdentifiers had different definitions of what a unqiue identifier was, so if you had a setting called 'setting', and clicked the setting dropdown on the next row, we'd watson because we'd attempt to add new setting called by the default "Setting". Treated them as the same, causes the new setting to be called "Setting1" and avoid the watson.