Make DependenciesSnapshot truly immutable
Previously the targets property was settable.

This commit removes all logic from the constructor (apart from basic
validation), making factory methods `FromChanges` and `RemoveTargets`
compute all state internally before calling the constructor, and
removing the private `MergeChanges` method.

A consequence of this is that when no changes occur, no allocation
takes place. The original object is returned unchanged.

Another consequence of this is that `anyChanges` is set to true
in the case that only the ActiveTarget property has changed, which
didn't previously happen.