Fix dependencies node regression
Fixes an error introduced in 77ca8fac8d423fbbfb1a43f5b1d71e0af3cb80f1
that caused many dependency tree nodes to appear with warning icons.

A unit test has been added to ensure this doesn't return.

The misconception in that earlier commit was that the Add method
replaces items that already present. There is no method on
ImmutableHashSet<>.Builder that achieves this, so we need the remove/add
combo.