Remove topLevelBuilder from IDependenciesSnapshotFilter
IDependenciesSnapshotFilter implementations are the source of a lot of
presentation logic for the dependencies node. Keeping this code correct,
tested and having good performance is a win all-round.

IDependenciesSnapshotFilter has a moderately complex and implicit
contract that's not enforced by the design, though I believe it could
be. This is a first step towards a design where implementations may
not violate the contract.

The 'top-level' dependencies collection is the subset of 'world'
dependencies where TopLevel is true.

Rather than require all filter implementations to correctly manage the
top-level collection separately as a HashSet<IDependency>, it's possible
to just compute that subset when needed directly from the world
dependencies.

Benchmarking shows this is faster overall, and allocates less memory.
The iteration/test is fast, and the in-memory representation becomes
an array rather than a persistent hash set.

The computation of HasUnresolvedDependency is now performed
during that same single scan, rather than lazily after the fact. The
previous check for this would scan both the top-level and world
dependencies, so this ends up being less work overall seeing as we're
scanning the world dependencies anyway.

The main benefit here is the guarantee of correctness. It's impossible
for the top-level and world dependency collections to contradict one
another.