Update meta analyzers in ruleset
dotnet/roslyn-analyzers#1861 has since been fixed, so the comment can be
removed. Rules in this group are meta-analyzers that apply to the source
code of analyzers themselves.

We do not have any analyzers in this repository so it's fine to disable these
diagnostics.

One is kept as it's of general use.