Remove change handling from IDependenciesGraphActionHandler
Of the two kinds of operations this interface supported, implementations
only provided one or the other.

Only one implementation provided the 'change update' handling, and that
behaviour has been inlined into DependenciesGraphChangeTracker directly
where it seems to be at home.

Methods on GraphActionHandlerBase now become abstract, as all subclasses
provide an implementation.

This removes a class, about 50 lines of code and simplifies the
abstraction.