Split IDependency and IDependencyModel
These types are not actually fungible and the inheritance only existed
to reuse the property declarations.

It's confusing/misleading to have inheritance here. For example, the Id
property on both types is different. For the model it's OriginalItemSpec,
while for the dependency it's a composition of provider type, target
framework and the model Id (OriginalItemSpec).

Similarly, IDependencyModel.DependencyIDs and IDependency.DependencyIDs
have different element formats.

This division makes this explicit, and allows for future alteration of
the internal IDependency interface.