Unresolved child test excludes non-visible children
These tests are used to determine whether a parent dependency tree node
should be displayed as unresolved. If an invisible child makes the parent
unresolved, it's confusing to the user.

For example, the NuGet node may be unresolved and the unresolved package
supplies an SDK, and is hidden from the tree. In such a case, the SDK
node will be marked as unresolved, and annotating the NuGet node is just
confusing and misleading.

See #4577.