Mark ProjectAssetsFileWatcher with "PackageReferences"
PackageRestoreInitiator was already marked with "PackageReferences" but the associated ProjectAssetsFileWatcher was still marked with ".NET".

Given PackageReferences is a dynamic capability, we needed to move over to implementing IProjectDynamicLoadComponent instead of ConfiguredProjectAutoLoad. AutoLoad can only be used for capabilities that are known at project creation time.

All the logic of ProjectAssetsFileWatcher is moved over to ProjectAssetsFileWatcherInstance, with Load renamed to InitializeAsync.