Add generated .vsconfig file
This file allows Visual Studio to verify it has the necessary components installed to work with the Project System solution.

The file was generated as described in this blog post:

https://devblogs.microsoft.com/setup/configure-visual-studio-across-your-organization-with-vsconfig/

These workloads were selected:

- .NET desktop development
- .NET Core cross-platform development
- Visual Studio extension development

The .vsconfig file contains the full list of components required to satisfy those workloads. We can probably remove several.