Change project reloader to actually factor in configs
ProjectReloaderIntecepter attempts to detect and prompt user to reload when changing from:

<Project>
  <PropertyGroup>
    <TargetFramework>net45</TargetFramework>
  </PropertyGroup>
</Project>

to

<Project>
  <PropertyGroup>
    <TargetFrameworks>net45;netcoreapp2.2</TargetFrameworks>
  </PropertyGroup>
</Project>

However, it did it in a way to cause it prompt in the following:

<Project>
  <PropertyGroup>
    <TargetFrameworkIdentifier>.NETFramework</TargetFrameworkIdentifier>
    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
  </PropertyGroup>
</Project>

to:

<Project>
  <PropertyGroup>
    <TargetFrameworks>net45</TargetFrameworks>
  </PropertyGroup>
</Project>

Despite these being equivalent.

Change the implementation to actually consult the dimension providers and avoid it having any knowledge how configurations are determined.