Fix path for test projects
Fix the path in the .editorconfig file used to refer to the test
projects. The test projects used to be under the src directory, in
subdirectories that all ended with "Tests". Now "tests" is a top-level
directory (a sibling of src) so we need to update the related path in
the .editorconfig file.

Given that all the test projects now have their own directory it would
make sense to split the test-specific settings in the .editorconfig into
a test-directory-specific .editorconfig. However, I'm going to leave
that to a future change rather than making this one even more complex.