Removes invalid Assumes.NotNull
This was added in 8f558fe5d4f6891164601017b48d15a7bdb83ea6 in an attempt to
suppress NRT warnings. The filePath variable may be null for non-project items,
and so this was triggering errors in VS, such as that reported in #5454.

Removing this line does not seem to introduce any new warnings, and fixes the
error.