Use strong types
The resource designer clearly predates the existence of generic types, as there are a number of places where it uses `Hashtable`, `ArrayList`, `ICollection`, `IList`, etc., rather than their generic equivalents. This requires we perform run-time type checks in a number of places, and makes the code harder to understand.

This change updates a number of places where we have a collection of `Resource`s.