Add ProjectImportsSubTreeProvider
This root graft tree provider adds an 'Imports' tree to the project that
shows all project imports.

Due to the size of this sub-tree, it is only displayed when the
per-project 'show all files' feature is enabled. When disabled, the tree
is removed from memory and its data subscription disposed.