Display import tree items in correct order
Previously, import tree items would always be displayed in alphabetical
order. With this change, their tree order reflects the actual order they
are evaluated in. This is helpful when tracing the value of a given
property, for example.

For CPS to respect the IProjectTree2.DisplayOrder property, the
SortByDisplayOrder capability must be present. This was only used in F#
projects before where it controlled both the display and manipulation of
item order. Consequently a new EditableDisplayOrder capability has been
added which covers the manipulation component, and has been added to
F#-specific commands/exports.