Use missing folder icon when Properties is missing from disk
Fixes: https://github.com/dotnet/project-system/issues/216

When the properties folder is explicitly included in the project when globs are turned off and source files are included from under it, or when explicitly included in the project via <Folder Include="Properties" /> and its missing from disk, use the missing folder icon to make it obvious as to why Add/Open Folder in File Explorer don't work.

Also made it easier for https://github.com/dotnet/project-system/issues/5803 to be implemented by plumbing out the expanded folder case.

Adding a specific missing folder icon for the App Designer is being tracked by: https://github.com/dotnet/project-system/issues/5803.