Don't leak old project instance
A project subscription (evaluation/dtb) can update in such a way that
produces no change to the dependencies tree.

Previously, when no tree changes were observed we would not trigger
an update. This, however, meant that we held a transitive reference to
the project instance (via the catalog collection).

This change ensures we always accept updates to other fields (not just
catalogs) regardless of whether IDependencyModel changes were made.