Add an Application property page for VS Online
Add a basic Application property page for VS Online. This includes _most_ of the functionality of the existing Application property page, with the missing parts called out at the end of the commit message. Most of the rest of this message describes how these changes work to produce the page.

This file is included in design-time builds run from VS. Here we specify that ApplicationPropertyPage.xaml is a `PropertyPageSchema` item to be used in the context of providing `Project` properties. Any such items in the `Project` context will be used to create project property pages. Currently ConfigurationGeneral.xaml and ProjectDebugger.xaml are _also_ part of the `Project` context--for reasons that aren't immediately clear to me--and so pages are created for them as well. They seem to be designed for programmatic access to properties rather than UI access, and aren't particularly functional at the moment.

The changes here ensure that ApplicationPropertyPage.xaml.cs is generated from ApplicationPropertyPage.xaml.

Here I'll call out some of the important parts of the new .xaml file and how they work.

Under `Rule.Categories` we define two categories: "General" and "Resources". The `Name` property uniquely identifies the category, while `DisplayName` is what is shown to the user in the UI (and will be translated into other languages). Individual properties can specify which category they belong to, which will affect how the UI controls are arranged.

The `Rule.DataSource` element specifies _defaults_ for how property information is retrieved. The `Persistence` attribute specifies what scheme is used for reading and writing the property values. In this case, the `ProjectFile` value means that the values are read directly from the project file itself, as opposed to the .user file or some other location. We'll see later how some individual properties use a different persistence scheme.

`SourceOfDefaultValue="AfterContext"` means that a default value for a property is specified in some .targets file imported at the end of the project file. The other option, `BeforeContext` would indicate that the default is specified in a .props file imported at the beginning of the project. This affects how we determine the default value to show for a property if it is not specified in the project file itself.

`HasConfigurationCondition="False"` means that the properties in the file are configuration-independent. If the use updates a property value, we will store the new value in a location in the project file that is used by all configurations rather than in one or more configuration-dependent property groups.

These are straight-forward strings, and the `Name` attribute indicates the MSBuild property they correspond to.

This is the most complicated property in the file as we're surfacing the `TargetFrameworkMoniker` property but if the user selects a different value we'll actually update the `TargetFramework` property. Let's first consider how we get the set of valid values that populate the drop down. The set of values is dynamically-determined, rather than being static values encoded in the .xaml file (see the `OutputType` property for an example of that). The `EnumProvider` attributes indicates that the values should come from the `SupportedTargetFrameworksEnumProvider`. This is handled by an existing type of the same name in the MS.VS.PS.Managed project. You can inspect its source for details, but it basically collects the `SupportedTargetFramework` items provided by the SDK and extracts the values and display names for those. These items basically look like this:

> `<SupportedTargetFramework Include=".NETCoreApp,Version=v3.1" DisplayName=".NET Core 3.1" />`

Note that the `Include` matches the format of the data in the `TargetFrameworkMoniker` property, which is how the property page finds the correct display name to show in the UI.

Setting the property to a new value is also interesting, as we don't want to update `TargetFrameworkMoniker` (it is a computed property) but we're not "connected" to the `TargetFramework` property, which we do want to update. To achieve this, we specify that the data source for this property uses "ProjectFileWithInterception" persistence. What this means is that when setting the value, it will first be routed through an `IInterceptingPropertyValueProvider`--specifically, the `TargetFrameworkMonikerValueProvider`. It in turns computes the short name for the given target framework moniker and explicitly sets the `TargetFramework` property.

Compared to `TargetFrameworkMoniker` this is simple. The output type can be one of three hard-coded values, each of which has a corresponding display name.

This is another instance where the set of values is computed dynamically, this time by the `StartupObjectsEnumProvider` type. It utilizes a Roslyn service to find valid entry points and returns those.

Of note here we set the `Subtype` attribute to indicate that the value is a file path. Eventually this will serve as an indicator that the UI should make a file picker available.

The `ApplicationManifest` property is interesting as the corresponding MSBuild property is just a file path, but there are a couple of special situations we need to handle. If a manifest isn't provided a default manifest can be created--we represent this as the "DefaultManifest" value. Alternatively, a developer can explicitly state they want no manifest. We represent this as the "NoManifest" value and it will cause us to set the `NoWin32Manifest` property in the project file. To handle these special cases we route getting and setting the value through the `ApplicationManifestValueProvider`.

This is a straight file path.

1. Framework multi-targeting. In the existing Application property page the "Target framework" control is disabled if the project file specifies multiple frameworks. I haven't found a way to dynamically mark a property as "read-only" yet, and so have not yet duplicated this behavior.
2. Suggested files. The existing Application page pre-populates the icons, manifest, and resources drop-downs with relevant files from the current project. A future change will add the dynamic value enum providers needed to support this.
3. No "Auto-generate binding redirects" checkbox. This property is only meaningful for projects targeting .NET Framework, and on the existing page we hide it when targeting .NET Core or Standard. As with read-only, I haven't yet figured out how to dynamically affect a property's visibility.