Add a Debug page with the current launch profile
This change adds a "Debug" property page with a single property: the active launch profile. There are a couple of distinct pieces worth explaining:

1. The "Debug" page itself. Currently this uses the "generic" page template rather than the "debugger" template. Pages utilizing the debugger template get some special handling in the property page system. Notably, the properties we want to show the user effectively all depend on the current profile, so if the profile changes we need to immediately refresh all of the other properties so they display the correct values. We also need support for adding and removing profiles via the page. The debugger page experience in VS Online does not currently support these things, and instead utilizes a single, hard-coded profile which will exist for C++ scenarios but not .NET scenarios. So for the time being we will stick with the generic page template in order to make progress on exposing the desired properties, though the page behavior will not always be correct (notably, if you change other values and _then_ change the profile without clicking "Apply" first you are likely to lose the changes, or end up with a non-sensical combination of settings).
2. The enum provider exposing the set of available profiles to populate the drop down. It turns out we already have this in the form of the `DebugProfileDebugTargetGenerator` and `DebugProfileEnumValuesGenerator` types. A later change will update these to support adding a new profile. Currently there is no UI support for that.
3. A means of getting and setting the active profile. Here I implement an `IInterceptingPropertyValueProvider` that gets called on attempts to get and set the "ActiveLaunchProfile" property. Instead of going back to the project file for the value, it goes to the `ILaunchSettingsProvider`.

An alternative to the `IInterceptingPropertyValueProvider` would be to create a whole new `IProjectPropertiesProvider` around the `ILaunchSettingsProvider` similar to the ones built into CPS for reading and writing *proj and *user files. The upside of the current approach is that it is much less code for the same effect. The downside is that `IInterceptingPropertyValueProvider` was originally created to support scenarios where property values read (or written to) the project file may need to be modified in some way, or where one property depends on another and they need to be updated together (see the `ApplicationManifestValuedProvider` for an example). As the debug settings will _always_ come from the `ILaunchSettingsProvider` rather than the project file there is a bit of a disjoint between how the system was intended to be used, and how it is actually being used.