Ignore problematic ILaunchSettingsUIProviders
In addition to requiring instatiation on the UI thread, some `ILaunchSettingsUIProvider`s also try to access UI constructs (like the application main window) in their constructors. These UI constructs may not exist in VS Online scenarios, and since we instantiate them lazily this can lead to exceptions when we're simply trying to get their supported launch commands.

Long-term we'll need to replace `ILaunchSettingsUIProvider` with something else that doesn't have a UI dependency. For the purposes of getting something working in VS Online, however, we just work around the issues for the moment.