Add new file system methods for getting file write times
We had races between checking that a file exists and reading its last write time, meaning that we could end up with DateTime.FromFileTime(0) in caches and fields.

Add two methods for the common usage in the tree, and treat situations where file cannot be accessed the same as the file not existing.