Don't collapse dependencies tree items on updates
Previously, changed nodes were removed from the tree and re-added, resulting in
different hierarchy IDs for the replacements and meaning that any expansion state
was not restored.

This logic existed to work around a limitation of the graph node implementation.

We no longer use graph nodes in the dependencies tree so can now fix this.