Remove ProjectTreeFlags.SupportsHierarchy
This flag indicated a node that could have graph node attached to it.
When present, tree construction would force the node to be removed
and re-added on resolved-state change, in order to trigger graph node
updates.

We no longer use graph nodes, and we no longer remove/add nodes on
updates.

This flag is public API. I cannot find usage within Microsoft code.