Implement dependencies tree search
Solution Explorer search occurs via the ISearchProvider interface. When a search is requested, we pass back tree items that match the search expression for the tree to display. At that point we are unlikely to have materialised items ready to return, so we create just those nodes that match. The tree then uses the "contained by" relationship to populate ancestor nodes, with the expectation that an ancestor will eventually exist in the current tree. For search, these are the IVsHierarchyItem objects that back top-level dependencies.

This bottom-up approach to building the tree required some new interfaces and data structures. Items that implement IContainedByAttachedItems can report their ancestors to the tree during search, meaning they can appear in search results. The snapshot now has the ability to lazily create a bottom-up dependency map for constant-time queries of a given item's ancestors.

Dependency tree item types now have factories for constructing their instances with a given directionality (relationship) in mind.

DependenciesSearchContext acts as a context object for a single search operation and monitors things like cancellation and ensuring maximum result limits are respected.