Prevent restore versions going back in time
Fixes: https://github.com/dotnet/project-system/issues/5904

We are merging data from configurations into a single value. This can cause data source versions to go back in time when switching from Debug -> Release -> Make some changes -> Debug.

We want to prevent that, so introduce a forward only block that will only publish new versions of a value when it has the same or higher version than the previous value. The expectation is that another value with a more recent data source version will appear shortly.